fix arquillian tests
