+++ dauphine-open-data/src/main/java/io/github/oliviercailloux/y2018/opendata/cas/Credentials.java
@@ -3,7 +3,6 @@ package io.github.oliviercailloux.y2018.opendata.cas;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 import javax.validation.constraints.NotNull;
 import javax.xml.bind.annotation.XmlAccessType;
